RoadkillPro for Maya 2018 Windows and Mac
Couple of bug fixes relating to Maintain UVs
New manual coming soon.
Links to Maths Kernel libs coming soon.

+++ Unwrapper/FindPins.cpp
@@ -202,7 +202,7 @@ void FindPins::SplitVerts()
 
 	while(iter != splitList.end())
 	{
-		int size = (*iter).second.size();
+		int size = (int)(*iter).second.size();
 
 		if(size > 1)
 		{
